보스 프론트 연결 끝~

+++ service/reviews.service.js
@@ -4,12 +4,13 @@ class ReviewService {
   reviewRepository = new ReviewRepository();
 
   //리뷰 생성
-  createReview = async (grade, comment, user_id, laundry_id) => {
+  createReview = async (grade, comment, user_id, laundry_id, boss_id) => {
     const createReviewData = await this.reviewRepository.createReview(
       grade,
       comment,
       user_id,
-      laundry_id
+      laundry_id,
+      boss_id
     );
 
     return createReviewData;
@@ -19,11 +20,15 @@ class ReviewService {
   };
 
   //리뷰 수정
-  updateReview = async (id, grade, comment) => {
+  updateReview = async (id, grade, comment, User) => {
     try {
       const findReview = await this.reviewRepository.findReviewById(id);
 
       if (!findReview) throw new Error("리뷰가 존재하지 않아요.");
+
+      if (User !== findReview.id) {
+        return {errormessage: "작성자가 아닙니다."}
+      }
 
       const updateReview = await this.reviewRepository.updateReview(
         id,
@@ -63,9 +68,13 @@ class ReviewService {
   };
 
   //리뷰 삭제
-  deleteReview = async (id) => {
+  deleteReview = async (id, User) => {
     const findReview = await this.reviewRepository.findReviewById(id);
     if (!findReview) throw new Error("리뷰가 존재하지 않아요.");
+
+    if (User !== findReview.id) {
+      return {errormessage: "작성자가 아닙니다."}
+    }
 
     await this.reviewRepository.deleteReview(id);
   };
